requirements.txt
discord.py[voice]
asyncpraw
colorama
motor
aiofiles
aiodns
aiohttp[speedups]
cchardet
pillow
humanize
psutil
uvloop
orjson
